docs: update disaster recovery guide for webhook
Since we moved to cert-manager for the webhook long time back, we
just need to disable webhook instead of backing the whole webhook.
And, after crs are restored we need to enable the webhook and all the
resources will be created itself by the operator.

Signed-off-by: subhamkrai <srai@redhat.com>
(cherry picked from commit 0dbfc66498f44b2629f03652b5fe0443cb994383)