docs: clarify in docs and readme that rook focuses on ceph
Rook has for quite some time only been focused on the Ceph
operator. Maintainers have only been focused on Ceph, and
the community has likewise been focused on Ceph. Now the docs
are updated to clarify that Ceph storage is the focus and
there are no plans for the project to support other storage
providers.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>