Improve POSIX compliance of local scripts
POSIX does not guarantee /bin/bash to exist. It mandates the presence of
`env`, however, which can be used to find bash.

This commit changes the shebang of local scripts to use /usr/bin/env to
improve platform compatibility of the toolchain.
It also adds lookups using `command -v` to replace any hard-coded /bin/bash
in makefiles.