test: change ceph version of osd on pvc canary test to avoid failures
This is to fix the following CI failure.

ci: Canary tests on PVCs are failing when creating OSDs #11600
https://github.com/rook/rook/pull/11600

We set the version of Ceph used for OSD on PVC to v16. It's the moving
target and was recently changed the target version. The new version
has a regression which results in the failure
of `ceph-volume raw prepare`. Although this will be fixed in Ceph
by the following PR, it's not backported to Quincy yet.

ceph-volume: fix a bug in lsblk_all() #49171
https://github.com/ceph/ceph/pull/49171

Let's change the version to v17.2.3, which is the latest version
without the above-mentioned regression of Ceph.

Signed-off-by: Satoru Takeuchi <satoru.takeuchi@gmail.com>