ci: split canary integration test in different files
having one huge file of canary tests in not helping us
in anyway instead it is making us look through the file and
importantly it making difficult to track the failing ci
which exact canary test is failing and other details. Currently,
this is very diffcult to trace at least for me.

Signed-off-by: subhamkrai <srai@redhat.com>