pool: add .mgr pool to the stretch cluster examples
The stretch clusters must create pools with a pool spec
that is valid for the stretch scenario, with 4 replicas
and failure domain and sub failure domain as needed
to match the topology. Otherwise, ceph will create a
default .mgr pool which will have replica 3 which is
invalid for the stretch scenario.

Signed-off-by: travisn <tnielsen@redhat.com>