ci: run object suite in master and release branches
The object suite was being skipped in master and release
branches. Since the object suite was only running in PRs
it was not noticed if there were any issues in the master
branch or release branches. Now we run it in these branches
in addition to the PRs.

Signed-off-by: travisn <tnielsen@redhat.com>
(cherry picked from commit 8fd46a46d3a1be8d032acf02591779476c58fd3d)