tests: start toolbox earlier in tests
The toolbox sometimes times out when the tests are waiting
for it to start. Now we create the toolbox spec sooner in
the tests so we won't need to wait so long for it to start
or increase the wait timeout in the tests.

Signed-off-by: travisn <tnielsen@redhat.com>
(cherry picked from commit 90d06a862b39a9e21c1d694aee72d191a4acc80a)