operator: add tolerations for rook-ceph operator deployment
This addresses #12660. As discussed in the issue when zone failure
happens and if the operator running in that zone, then it takes the
default toleration of 300 seconds. We want to override this default
toleration seconds to 5 seconds so that operator gets redeployed
immediately

Signed-off-by: Mahesh Shetty <mashetty@redhat.com>
(cherry picked from commit 518537113b0310e6e5c38ccecf5a34f753f6ba63)