build: update to latest libopenstorage/secrets lib
Update to the latest version of libopenstorage/secrets which does not
have a dependency on Hashicorp Vault libraries that are licensed under
the Business Source License (BUSL-1.1). The latest version of the
library has been modified to use Vault's open source API that is
licensed under Mozilla Public License (MPL-2.0).

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>