security: enhance SCC Function to Support Multiple Namespaces
Currently, the function receives a single namespace string and appends
it to the users. This approach is insufficient when dealing with
multiple CephClusters running in different namespaces, as there will be
only one SCC. This update modifies the function to create users for all
specified namespaces, ensuring compatibility with multiple namespaces.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>