helm: add namespace to all resource templates
The helm templates assumed that the resources would be installed
to the given namespace for the helm install or upgrade. This works
perfectly until there is a desire to extract the manifests from
the helm chart and instead install with those. Thus, the namespace
is added to all the resources in the chart where they were
missing.

Signed-off-by: travisn <tnielsen@redhat.com>
(cherry picked from commit b4e51f6d12bcc2bcb6a8293a522add5acbb4f367)