core: increase the ceph command timeouts
We have seen behaviors in which the timeout function
signal killing destroys the operator flow.
For example:
I | exec: exec timeout waiting for process radosgw-admin
to return. Sending interrupt signal to the process
I | ceph-object-controller: multisite configuration
for object-store ocs-storagecluster-cephobjectstore is complete
Instead of giving the error to the object controller,
 that command was successful it actually broke the operator
flow and print success
So to avoid these flows we should give commands more time to execute

Signed-off-by: parth-gr <partharora1010@gmail.com>