object: always proxy radosgw-admin commands via mgr pod
It is an ongoing difficulty to keep up with the Ceph version
incompatibilities where radosgw-admin is involved. The tool's developers
seem to do a good job of making it backwards compatible but not forwards
compatible. This means that any time a new Ceph version introduces a
substantial change to radosgw-admin, Rook must update its base image to
the latest Ceph version in order to support the newest Ceph version.

Sometimes, Rook is unable to upgrade to a new Ceph version due to a bug,
which means that there then can be critical incompatibilities.

To try to resolve these incompatibilities, let's permanently enable one
of the Multus code paths: the "command proxy" sidecar in the mgr pod.
The mgr pod should always be running the same Ceph image version as the
RGW pod, or it may be running a higher version if the cluster is in the
middle of being upgraded. In either case, there should be no issues with
radosgw-admin version mismatches.

There may be radosgw-admin version mismatch issues if the cluster is
ever downgraded; however, this risk also exists today. I don't believe
this risk can be mitigated without much deeper architectural discussion,
and I don't believe it is necessary since we do not officially support
version downgrades.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>