add cert-manager and trust-manager
The keystone test cases lack the cert-manager installation and fail
with error (on creating CRs which are not available without the
cert-manager CRDs). The ca certificate file of the self signed ca
is not available to the CephObjectStore deployment, but is necesary
to contact keystone using tls.

This commit installs cert-manager and trust-manager using helm.
Additionally a ca bundle is generated by trust-manager to create
configmaps with a ca-bundle file which includes the self-signed ca
certificate. This can later be used by the CephObjectStore deployment
to create a tls connection with keystone.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>