configurable ca certs configmap for CephObjectStore
The CephObjectStore deployment uses tls to connect to
keystone. This can involve a selfsigned ca certificate.
Currently it is not possible to configure the ca certificate
files used to verify the tls certificates.

This commit adds a new spec entry "caConfigMapName" to
provide a name of a config map. This config map contains
a ca-bundle.crt file (e.g. provided by trust-manager)
which contains the ca certificates to use on verifying a
tls certificate. The config map is included as a volumeMount
in the CephObjectStore deployment.

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>

Revert "fixup! configurable ca certs configmap for CephObjectStore"

This reverts commit 4c29f13f7eed1afe858c60bbe81cd92f6be5b3c8.

Revert "fixup! configurable ca certs configmap for CephObjectStore"

This reverts commit 4f18653509f2c4d30d11628567034922a20c1403.

Revert "configurable ca certs configmap for CephObjectStore"

This reverts commit e168b30979a4956c5e0496559ebb300e42e4cc5c.